Fixed an error in logout
This simple key lookup error was discovered while fiddling around with a
behave feature. This highlights a weakness in the current testing
regime: the views are not all touched by a test, leading to simple
errors in endpoints that have only framework-integrated logic like this
one. The solution is to test the application with mock user application,
without a database connection and the whole stack. Will get back to this
later.