Added company table, application
Before introducing storage locations and per-company products, in
addition to enabling inviting other users, I decided to refactor the
company field on the user to a separate table. A new application would
be responsible for this database entity, presenting domain logic for the
table to the views and other applications. In this step, it is possible
to observe the effect of which tests to focus on in order to carry out
the refactoring. The choice is between the behave features and unit
tests. I chose the former, changing code in one refactoring step,
because I thought it would be something small, and then modifying the
unit tests to fit the change. This turned out to be more difficult than
expected, and probably the wrong choice.

The difficulties with behave features stem from two things: Difficulty
of debugging and understanding features when they fail, and loss of
focus on the small pieces. When a feature fails, one has to debug the
whole application, setting break points in various places; this is not
the case with unit tests. At this scale, the break point approach works,
but is still more painful than necessary, and will become even more so
when code grows. The loss of focus derives from the need to just "get it
to work", instead of concentrating on individual interface methods and
details. With unit tests, one has more to fix when refactoring, but the
detailed view is a great plus. Also, there is a bug in the code built
using features, where a new company is created even when the user
creation is not successful. This bug would have been caught if
development were done with unit tests.