Implemented saving an invitation, test
One advantage of the chosen approach can be observed in this step: The
persistance of a row in database table can be implemented and unit
tested totally independently of the table itself. Thus, one does not
have to start with an untested table and then build the code around
it. Instead, the behaviro of the test is asserted in the unit tests, and
it can be used a specification in the models.