Added behave to spec stock intake
For the stock feature, I wanted to start from the API side. Using
behave, we can create a specification for the API, and then develop the
details from there. The behave test drives the initial code, but then as
things get more intricate, we can resort to finer design and unit
tests.