Decorator to check user is logged in
The views are getting cluttered, and I would like to stick to the
'functional core, imperative shell' method, which means that some more
space is needed to put that shell without making the controllers too
big. This decorator pulls out the user check that is orthogonal to
functionality, and puts it around the view function. Another thing while
developing this decorator was that I could experiment and didn't need to
write these unit tests, thanks to the integration tests. Once I knew the
idea was working I could develop it in a proper manner, and also uni
test it.