Cleanup of stock intake unit tests
The stock intake code is too imperative for my taste, with a loop on
product input that should filter out errors in the intake process. This
is more convoluted to test then a simple map or similar to returns the
diff for the database instead of changing state in the loop. If the
whole loop construct were to be tested, there would be a ton of state
coordination for the tests. I don't have a solution for this right now,
however, and suspect that the DB interface should be something other than
an ORM so that stat eis not tied to data representation. The best I could
do for the time being is to locate the updating of stock in a relatively
simple function so that it can be tested in isolation.