Test for intake from product list
This test is rather long, since it has to set up different kinds of data
to test for the various conditions. Because arranging the needed DB
state is so annoying, I put three conditions into one test, since the
details are handled in another test. I would like to refactor the code
and the test, however, and make it smarter.