Created Stock model
I really like the fact that the stock model can be created in the way
expected by the code written with unit tests, and it works. The only
problem was that there was an issue with the default values in the Stock
model, in that they were being used before being set.