Method to get stock for storage
The product application is getting unnecessarily big. In the next
iteration on stock features, the stock-related stuff will definitely get
moved to its own application. One thing I really like about testing
with the application objects, as seen in this test, is that the
application can be instantiated with a None for tables or other
applications that should not be accessed. This is a very easy and safe
way to assert that no queries on these entities will be made.